Update TopologyManagerBellmanFord.java
Floodlight uses a topology manager to maintain state of all connected switches and nodes. This helps the controller to forward requests from one source to destination. Hence the topology manager is primarily used by the Forwarding module.
It considers each node as the root of a cluster. Hence there are as many clusters as there are nodes. For each cluster, the shortest path broadcast tree is computed. By default floodlight considers the link cost for each hop as 1. There is no ideal way to set this link cost with floodlight API or provide values from a simulated environment like that of Mininet.
The shortest path within a cluster, starting from root node to each of the other nodes is computed using Dijkstra algorithm. This algorithm is efficient for local routing when we are aware of all connections. We have made an attempt to remove Dijkstra algorithm and replace the core module with Bellman Ford algorithm for two main reasons.
1.	Bellman Ford algorithm allows negative link costs provided the network does not have any loops with negative link costs. Note: Floodlight today checks for loops but does not count costs. This feature can leverage applications that want to assign link costs for bandwidth or performance tuning and also aid in decreasing link costs.
2.	The core library can be extended in future to implement any distance vector routing algorithm with message passing on link failures or partial view of the network. This will make use of Distributed Bellman Ford algorithm